reddit_api/requirements.txt
praw
python-dotenv
pytest
ruff
black